Changes to the bypass-sign-change-blocking setting.
The setting is enabled by default now since we usually want to always bypass protection plugins such as WorldGuard, which reacts to sign editing by default.
If the setting is enabled, we now cancel the sign change event early so that other plugins (such as WorldGuard) ignore it and don't print their cancellation message.